Created `util::integrate_matrix' function.
Moved the integration of matrices to the `gauss_quadrature.h' header file.
Different matrix formulations use different integrals but the operation to
integrate each is the same, and thus it makes more sense to place that function
within the Gaussian quadrature header.  Function accepts a functor called
``Matrix_Func'' which reports the size of the final matrix and loops and
integrates over each coefficient in the matrix.